refactor proof pack detail router
